Refresh balances when bonding shares change
