+++ components/common/utils.ts
@@ -1,4 +1,5 @@
-import { ethers, ContractTransaction } from "ethers";
+import { ethers, ContractTransaction, BigNumber } from "ethers";
+import { ERC1155Ubiquity } from "../../contracts/artifacts/types";
 import { useEffect } from "react";
 
 export function logGas(txDone: ethers.ContractReceipt) {
@@ -23,4 +24,20 @@ export async function performTransaction(transaction: Promise<ContractTransactio
     console.error("Transaction error", e);
   }
   return false;
+}
+
+export async function erc1155BalanceOf(addr: string, erc1155UbiquityCtr: ERC1155Ubiquity): Promise<BigNumber> {
+  const treasuryIds = await erc1155UbiquityCtr.holderTokens(addr);
+
+  const balanceOfs = treasuryIds.map((id) => {
+    return erc1155UbiquityCtr.balanceOf(addr, id);
+  });
+  const balances = await Promise.all(balanceOfs);
+  let fullBalance = BigNumber.from(0);
+  if (balances.length > 0) {
+    fullBalance = balances.reduce((prev, cur) => {
+      return prev.add(cur);
+    });
+  }
+  return fullBalance;
 }
